Demonstrating the tricky runWithDefault example
I copied the initial example from Slack and the runWithDefault proposed
implementation.

Seeing the error as described, I tried to narrow down the type error. To
do this I started by relying on the inferred type for decoder, assuming
it was Decoder<User> (though it obviously isn't, I'm trying to find a
more specific contradiction). Then I simply used the new
`runWithDefault` method, making sure to state that I'm expecting a User.
I gave a default value and a sample JSON string and Typescript gave the
following error:

```
[!] (rpt2 plugin) Error: /home/dwilson/json-type-validation/src/example.ts(14,43): semantic error TS2345 Argument of type '{ id: number; }' is not assignable to parameter of type '{ id: number; isDog: boolean | undefined; }'.
  Property 'isDog' is missing in type '{ id: number; }'.
```

This suggests to me that the decoder as defined is for an object which
must have an `isDog` key, that might be undefined. Our User object is
slightly different from Typescripts perspective, in that it may or may
not have an isDog key that must be a boolean if present.

This suggests the problem is in the interaction between object and
optional.